version 0.0.1
in v0.0.1, js files were updated to display words and delete them, to also change which mode is active and be able to delete the numbers list. css was updated to make index.html look more presentable. index.html was updated to have a better markup and be able to adjust to the new words